MapleWebBackEnd/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cors==1.0.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
idna==3.10
mysqlclient==2.2.7
pycparser==2.23
PyJWT==2.10.1
PySocks==1.7.1
python-decouple==3.8
requests==2.32.5
requests-file==2.1.0
sqlparse==0.5.3
tldextract==5.3.0
typing_extensions==4.15.0
urllib3==2.5.0
zope.event==6.0
zope.interface==8.0
